+++ app/Interfaces/CustomerProfileServiceInterface.php
@@ -1,0 +1,6 @@
+<?php 
+namespace App\Interfaces;
+interface CustomerProfileServiceInterface
+{
+
+}
